EduHub Version 2 @Tahmid_Chy
This text describes a website called EduHub, which is focused on online learning and education. The author of the text mentions that they have worked on the front-end development of the website using HTML and CSS3. Specifically, they have coded the front end based on a PSD format. The author acknowledges that there is room for improvement in their work.